shopstop_backend/requirements.txt
asgiref==3.2.3
coverage==5.0.3
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
Markdown==3.2
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
PySocks==1.7.1
pytz==2019.3
sqlparse==0.3.0
uWSGI==2.0.18
